Retire batch.schedule as storage; derive the schedule live
Epic #208 set out to make schedule/equipment/gravity "pure views derived live
from the brewable". Equipment (#210) and gravity (#211) got there; the schedule
did not. #212 moved its last user-owned fields (checkoff/actual/pitch date) onto
the tracker, which left `batch.schedule` a persisted cache of a pure function of
`brewable` — with exactly one reader, sitting in the same useMemo where
equipment and gravity were already derived live.

- actions/_updateSchedule.ts -> actions/deriveSchedule.ts: now a pure
  `deriveSchedule(brewable) => ScheduleItem[]`. The previous/sameDerived
  reuse-by-reference map is gone — it existed only to keep the persisted array
  reference-stable for updateBatch's isEqual; a memo needs none of that.
- model/batch.ts: drop the `schedule` field. ScheduleItem/ScheduleDetail stay
  (still the derived shape).
- createBatch/updateBatch: no longer project a schedule. Shopping is now the
  only persisted derivation (it alone carries user-owned cost/purchased).
- screen/batch-schedule: derives in the existing panels useMemo, deps
  [data.phases, data.brewable].

Fixes a latent bug class as a side effect: a batch whose stored schedule was
built by older code kept dead write-through paths, so a brew-day edit wrote a
removed field and was silently lost. Deriving live regenerates paths from the
current brewable every time, so that can't happen.
